login/Register + authentication (adding/edit/delete listing/review)

+++ 47. Major Project/controllers/listingRoutes.js
@@ -5,6 +5,8 @@ const ExpressError = require("../utils/ExpressErrors");
 const listingRouter = require("express").Router();
 const { listingSchema } = require("../Schema.js");
 
+const checkAndHanldeNotLoggedIn = require("../utils/module.js");
+
 const validateListing = (req, res, next) => {
   const result = listingSchema.validate({ listing: req.body });
   if (result.error) {
@@ -21,44 +23,11 @@ listingRouter.get(
 );
 
 listingRouter.get(
-  "/listings",
-  wrapAsync(async (req, res) => {
-    const allListings = await Listing.find({}).populate("reviews");
-    res.send(allListings);
-  })
-);
-
-listingRouter.post(
-  "/",
-  // validateListing,
-  wrapAsync(async (req, res, next) => {
-    // if(!req.body.title && !req.body.description && !req.body.image && !req.body.price && !req.body.location && !req.body.country){}
-    const newListing = new Listing(req.body);
-    await newListing
-      .save()
-      .then((res) => {
-        req.flash("success", {
-          msg: `'${res.title}' Listing Created Successfully`,
-          type: "normal",
-        });
-      })
-      .catch((err) => {
-        req.flash("failure", err);
-      });
-    res.redirect("/");
-  })
-);
-
-listingRouter.get(
   "/:id",
   wrapAsync(async (req, res) => {
     const listing = await Listing.findById(req.params.id).populate("reviews");
-    console.log("listing", listing);
     if (!listing) {
-      req.flash("failure", {
-        msg: `Listing not found`,
-        type: "notFound",
-      });
+      req.flash("error", `Listing not found`);
       res.redirect("/");
     } else {
       let netRating = listing.reviews.map((i) => i.rating);
@@ -72,15 +41,36 @@ listingRouter.get(
   })
 );
 
+listingRouter.post(
+  "/",
+  // validateListing,
+  wrapAsync(async (req, res, next) => {
+    if(checkAndHanldeNotLoggedIn(req,res,`You must be logged in to add a new listing`)){
+      return;
+    }
+    // if(!req.body.title && !req.body.description && !req.body.image && !req.body.price && !req.body.location && !req.body.country){}
+    const newListing = new Listing(req.body);
+    await newListing
+      .save()
+      .then((res) => {
+        req.flash("success", `'${res.title}' Listing Created Successfully`);
+      })
+      .catch((err) => {
+        req.flash("error", err);
+      });
+    res.redirect("/");
+  })
+);
+
 listingRouter.get(
   "/edit/:id",
   wrapAsync(async (req, res) => {
+    if(checkAndHanldeNotLoggedIn(req,res,`You must be logged in to edit a listing`)){
+      return;
+    }
     const listing = await Listing.findById(req.params.id);
     if (!listing) {
-      req.flash("failure", {
-        msg: `Listing not found`,
-        type: "notFound",
-      });
+      req.flash("error", `Listing not found`);
       res.redirect("/");
     } else {
       res.render("listings/edit", { place: listing });
@@ -92,24 +82,15 @@ listingRouter.patch(
   "/:id",
   // validateListing,
   wrapAsync(async (req, res) => {
-    // try{
-    //     await Listing.findByIdAndUpdate( req.params.id , {$set: req.body})
-    //     res.redirect(`/listings`)
-    // } catch(err){
-    //     next(err)
-    // }
-    // if(!req.body.title || !req.body.description || !req.body.image || !req.body.price || !req.body.location || !req.body.country){
-    //     throw new ExpressError(400,"Recieved No Updation Data")
-    // }
+    if(checkAndHanldeNotLoggedIn(req,res,`You must be logged in to edit a listing`)){
+      return;
+    }
     await Listing.findByIdAndUpdate(req.params.id, { $set: req.body })
       .then((res) => {
-        req.flash("success", {
-          msg: `'${res.title}' Listing Updated Successfully`,
-          type: "normal",
-        });
+        req.flash("success", `'${res.title}' Listing Updated Successfully`);
       })
       .catch((err) => {
-        req.flash("failure", err);
+        req.flash("error", err);
       });
     res.redirect(`/${req.params.id}`);
   })
@@ -118,15 +99,15 @@ listingRouter.patch(
 listingRouter.delete(
   "/:id",
   wrapAsync(async (req, res) => {
+    if(checkAndHanldeNotLoggedIn(req,res,`You must be logged in to delete a listing`)){
+      return;
+    }
     await Listing.findByIdAndDelete(req.params.id)
       .then((res) => {
-        req.flash("success", {
-          msg: `'${res.title}' Listing Deleted Successfully`,
-          type: "delete",
-        });
+        req.flash("success", `'${res.title}' Listing Deleted Successfully`);
       })
       .catch((err) => {
-        req.flash("failure", err);
+        req.flash("error", err);
       });
     // await Review.deleteMany({listing:req.params.id})
     // instead of the above action we define a post mongoose middleware (check it in models/listing.js)
@@ -137,6 +118,9 @@ listingRouter.delete(
 listingRouter.get(
   "/addNew/new",
   wrapAsync(async (req, res) => {
+    if(checkAndHanldeNotLoggedIn(req,res,`You must be logged in to add a new listing`)){
+      return;
+    }
     res.render("listings/newListing");
   })
 );
